祝日管理完成

+++ abe/holiday_app/holiday/views/maintenance_data.py
@@ -3,9 +3,10 @@ from holiday import app
 from functools import wraps
 from holiday.models.mst_holiday import Entry
 from holiday import db
+from flask_sqlalchemy import SQLAlchemy
 #from flask_blog.views.views import login_required
 
-@app.route("/maintenance_date",methods=["GET","POST"])
+@app.route("/maintenance_data",methods=["GET","POST"])
 def maintenance():
     if request.method=="POST":
         if request.form["button"]=="insert_update":
@@ -27,12 +28,29 @@ def maintenance():
                 return render_template("result.html",Date=Date,Text="("+Text+")"+"が登録されました")
         
         elif request.form["button"]=="delete":
-            entry=Entry.query.get(request.form["holiday"])
-            db.session.delete(entry)
-            db.session.commit()
-            return redirect(url_for("show_entries"))
-        
 
+            if request.form["holiday"]=="":
+                flash("日付が未入力です。入力してください")
+                return redirect("/")
+    
+            else:
+                Date=request.form["holiday"]
+                entry=Entry.query.get(request.form["holiday"])
+                print(entry)
+
+                if entry != None:
+                    #entry_text=Entry.query.get()
+                    #Text=Entry.query.get(entry.holi_text)
+                    Text=entry.holi_text
+                    #print(Text)
+                    db.session.delete(entry)
+                    db.session.commit()
+                    return render_template("result.html",Date=Date,Text="("+Text+")"+"は削除されました")
+                
+                else:
+                    flash(f"{Date}は、祝日マスタに登録されていません")
+                    return redirect("/")
+                            
 @app.route("/result",methods=["GET","POST"])
 def result():
     return redirect(url_for("show_entries"))
